Revert previous commit.
The problem is that the value of the new key in the dict should likely not be a dict but a dataclass.
And that dataclass may have default values that would not be set if we dynamically create a new dict value.
The only way to resolve this is to make sure that in the pvml global config the key is already set (so a default dataclass is generated) after which we can override any properties using the update() function.